+++ elmyra.ip.access.epo/CHANGES.rst
@@ -5,6 +5,7 @@ elmyra.ip.access.epo CHANGES
 development
 ===========
 - middleware: api endpoint for downloading a zip archive of multiple pdf documents
+- ui: wire multiple pdf zip archive download
 
 0.39.0
 ======